LayoutTests/media/modern-media-controls/layout-node/visible-expected.txt

Testing the LayoutNode.visible property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking default value
PASS node.visible is true
PASS node.element.style.display is ""

node.visible = false
PASS node.visible is false
PASS node.element.style.display is ""

Layout was performed
PASS node.visible is false
PASS node.element.style.display is "none"

node.visible = true

Layout was performed
PASS node.visible is true
PASS node.element.style.display is "inherit"

node.visible = true
PASS scheduler.hasScheduledLayoutCallbacks is false

PASS successfullyParsed is true

TEST COMPLETE
